Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main(){
     float average;
-    int a,b,c;
+    float a,b,c;
     printf("");
-    scanf("%d %d %d",&a,&b,&c);
+    scanf("%.2f %.2f %.2f",&a,&b,&c);
     average = (a+b+c)/3;
     printf("Average: %.2f\n",average);
     return 0;
